Fix memory leak in DirectWrite backend
This fixes a memory leak due to a missing call to release
on the localized names in populateFamilyAliases(). The
DirectWriteScope makes this automatic when it goes out
of scope.

We did release in the other places, but for hardening
the code a bit, this also adjusts them to use the scope
guard.

Change-Id: I88402fad930e36cdd3a78244941fb53ca214520e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>